+++ src/SqlClient.Tests/ProgrammabilityTests.fs
@@ -1,10 +1,10 @@
 ﻿module FSharp.Data.ProgrammabilityTest
 
+open System.Data.SqlClient
 open Xunit
 open FsUnit.Xunit
 
 [<Literal>] 
-//let connection = ConnectionStrings.AdventureWorksAzure
 let connection = ConnectionStrings.AdventureWorksNamed
 
 type AdventureWorks = SqlProgrammabilityProvider<connection>
@@ -28,6 +28,13 @@ let ScalarValuedFunction() =
     //async execution
     Assert.Equal(Some(sprintf "%08i" x), cmd.AsyncExecute(x) |> Async.RunSynchronously)
 
+[<Fact>]
+let ConnectionObject() =
+    let conn = new SqlConnection(ConnectionStrings.AdventureWorksLiteral)
+    let cmd = new GetLeadingZeros()
+    let x = 42
+    Assert.Equal( Some(sprintf "%08i" x), cmd.Execute(x))
+
 type Address_GetAddressBySpatialLocation = AdventureWorks.Person.Address_GetAddressBySpatialLocation
 open Microsoft.SqlServer.Types
 
@@ -36,3 +43,6 @@ let ``GEOMETRY and GEOGRAPHY sp params``() =
     use cmd = new Address_GetAddressBySpatialLocation()
     cmd.AsyncExecute(SqlGeography.Null) |> ignore
     
+[<Fact>]
+let routineCommandTypeTag() = 
+    Assert.Equal<string>(ConnectionStrings.AdventureWorksNamed, GetContactInformation.ConnectionStringOrName)